Insert and remove nodes with child modifications
